+++ CHANGELOG.md
@@ -5,6 +5,21 @@ This changelog is according to [Keep a Changelog](http://keepachangelog.com).
 All notable changes to this project will be documented in this file.
 We will follow [Semantic Versioning](http://semver.org/) from version 2 and onwards.
 
+## 1.30.0 March 5th, 2018
+### Added
+* Adds a setter for titleWidth to the snippet preview. 
+* Adds a researcher to calculate the reading time for a given paper.
+* Adds a filter to mark Spanish sentences as non-passive when certain exception words occur between the auxiliary and the participle. The list of exception words includes all forms of the copula 'estar'.
+* Adds transition words assessment for Portuguese, props [amesdigital](https://github.com/amesdigital).
+* Adds prominent words for Portuguese, props [amesdigital](https://github.com/amesdigital).
+
+### Changed
+* The snippet preview now shows the mobile preview by default.
+* Fixes a bug where division by zero errors in the passive voice assessment would cause `NaN%` to show up in the feedback.
+* Fixes a bug where multiple `rel` arguments prevented correct `nofollow` detection. 
+* Slightly increased the height of the meta description box so it matches the maximum amount of characters without needing a scrollbar.
+* Improves the list of Portuguese function words.
+
 ## 1.29.0 January 15th, 2018
 ### Added
 * Adds language support table to the README.